+++ src/components/mod.rs
@@ -1,2 +1,8 @@
 pub mod result;
 pub mod test;
+
+#[derive(Debug, Eq, PartialEq, Clone, Hash)]
+pub enum Screen {
+    Test,
+    Results,
+}
