+++ src/components/test/spans.rs
@@ -1,0 +1,45 @@
+// Helpers
+
+fn words_to_spans<'a>(
+    words: &'a [TestWord],
+    current_word: usize,
+    theme: &'a Theme,
+) -> Vec<Vec<Span<'a>>> {
+    let mut spans = Vec::new();
+
+    for word in &words[..current_word] {
+        let parts = split_typed_word(word);
+        spans.push(word_parts_to_spans(parts, theme));
+    }
+
+    if current_word < words.len() {
+        let parts_current = split_current_word(&words[current_word]);
+        spans.push(word_parts_to_spans(parts_current, theme));
+
+        for word in &words[current_word + 1..] {
+            let parts = vec![(word.text.clone(), Status::Untyped)];
+            spans.push(word_parts_to_spans(parts, theme));
+        }
+    }
+    spans
+}
+
+fn word_parts_to_spans(parts: Vec<(String, Status)>, theme: &Theme) -> Vec<Span<'_>> {
+    let mut spans = Vec::new();
+    for (text, status) in parts {
+        let style = match status {
+            Status::Correct => theme.prompt_correct,
+            Status::Incorrect => theme.prompt_incorrect,
+            Status::Untyped => theme.prompt_untyped,
+            Status::CurrentUntyped => theme.prompt_current_untyped,
+            Status::CurrentCorrect => theme.prompt_current_correct,
+            Status::CurrentIncorrect => theme.prompt_current_incorrect,
+            Status::Cursor => theme.prompt_current_untyped.patch(theme.prompt_cursor),
+            Status::Overtyped => theme.prompt_incorrect,
+        };
+
+        spans.push(Span::styled(text, style));
+    }
+    spans.push(Span::styled(" ", theme.prompt_untyped));
+    spans
+}
