+++ src/components/test/status.rs
@@ -1,0 +1,92 @@
+#[derive(PartialEq, Clone, Copy, Debug)]
+enum Status {
+    Correct,
+    Incorrect,
+    CurrentUntyped,
+    CurrentCorrect,
+    CurrentIncorrect,
+    Cursor,
+    Untyped,
+    Overtyped,
+}
+
+fn split_current_word(word: &TestWord) -> Vec<(String, Status)> {
+    let mut parts = Vec::new();
+    let mut cur_string = String::new();
+    let mut cur_status = Status::Untyped;
+
+    let mut progress = word.progress.chars();
+    for tc in word.text.chars() {
+        let p = progress.next();
+        let status = match p {
+            None => Status::CurrentUntyped,
+            Some(c) => match c {
+                c if c == tc => Status::CurrentCorrect,
+                _ => Status::CurrentIncorrect,
+            },
+        };
+
+        if status == cur_status {
+            cur_string.push(tc);
+        } else {
+            if !cur_string.is_empty() {
+                parts.push((cur_string, cur_status));
+                cur_string = String::new();
+            }
+            cur_string.push(tc);
+            cur_status = status;
+
+            // first currentuntyped is cursor
+            if status == Status::CurrentUntyped {
+                parts.push((cur_string, Status::Cursor));
+                cur_string = String::new();
+            }
+        }
+    }
+    if !cur_string.is_empty() {
+        parts.push((cur_string, cur_status));
+    }
+    let overtyped = progress.collect::<String>();
+    if !overtyped.is_empty() {
+        parts.push((overtyped, Status::Overtyped));
+    }
+    parts
+}
+
+fn split_typed_word(word: &TestWord) -> Vec<(String, Status)> {
+    let mut parts = Vec::new();
+    let mut cur_string = String::new();
+    let mut cur_status = Status::Untyped;
+
+    let mut progress = word.progress.chars();
+    for tc in word.text.chars() {
+        let p = progress.next();
+        let status = match p {
+            None => Status::Untyped,
+            Some(c) => match c {
+                c if c == tc => Status::Correct,
+                _ => Status::Incorrect,
+            },
+        };
+
+        if status == cur_status {
+            cur_string.push(tc);
+        } else {
+            if !cur_string.is_empty() {
+                parts.push((cur_string, cur_status));
+                cur_string = String::new();
+            }
+            cur_string.push(tc);
+            cur_status = status;
+        }
+    }
+    if !cur_string.is_empty() {
+        parts.push((cur_string, cur_status));
+    }
+
+    let overtyped = progress.collect::<String>();
+    if !overtyped.is_empty() {
+        parts.push((overtyped, Status::Overtyped));
+    }
+    parts
+}
