--- src/types/fraction.rs
@@ -1,47 +1,0 @@
-use std::cmp;
-use std::fmt;
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Default)]
-pub struct Fraction {
-    pub numerator: usize,
-    pub denominator: usize,
-}
-
-impl Fraction {
-    pub fn new(numerator: usize, denominator: usize) -> Self {
-        Self {
-            numerator,
-            denominator,
-        }
-    }
-}
-
-impl From<Fraction> for f64 {
-    fn from(fraction: Fraction) -> Self {
-        if fraction.denominator == 0 {
-            0.0
-        } else {
-            fraction.numerator as f64 / fraction.denominator as f64
-        }
-    }
-}
-
-impl cmp::Ord for Fraction {
-    fn cmp(&self, other: &Self) -> cmp::Ordering {
-        f64::from(*self)
-            .partial_cmp(&f64::from(*other))
-            .unwrap_or(cmp::Ordering::Equal)
-    }
-}
-
-impl cmp::PartialOrd for Fraction {
-    fn partial_cmp(&self, other: &Self) -> Option<cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl fmt::Display for Fraction {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}/{}", self.numerator, self.denominator)
-    }
-}
