--- src/types/test/mod.rs
@@ -1,30 +1,0 @@
-pub mod handler;
-use crate::types::TestWord;
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct Test {
-    pub words: Vec<TestWord>,
-    pub current_word: usize,
-    pub complete: bool,
-    pub backtracking_enabled: bool,
-    pub sudden_death_enabled: bool,
-    pub backspace_enabled: bool,
-}
-
-impl Test {
-    pub fn new(
-        words: Vec<String>,
-        backtracking_enabled: bool,
-        sudden_death_enabled: bool,
-        backspace_enabled: bool,
-    ) -> Self {
-        Self {
-            words: words.into_iter().map(TestWord::from).collect(),
-            current_word: 0,
-            complete: false,
-            backtracking_enabled,
-            sudden_death_enabled,
-            backspace_enabled,
-        }
-    }
-}
