--- src/types/timing_data.rs
@@ -1,10 +1,0 @@
-use std::collections::HashMap;
-use tuirealm::ratatui::crossterm::event::KeyEvent;
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct TimingData {
-    // Instead of storing WPM, we store CPS (clicks per second)
-    pub overall_cps: f64,
-    pub per_event: Vec<f64>,
-    pub per_key: HashMap<KeyEvent, f64>,
-}
